Phase 1+2: RDS bit-transition sync detector + pipeline integration
Replaces the FM pilot zero-crossing sync (which had unresolvable N*52.6 usec
ambiguity across nodes) with an RDS bit-boundary sync.  All nodes receiving
the same FM station see the same RDS bit transitions, eliminating the
ambiguity and giving sub-microsecond TDOA precision.

New module: src/beagle_node/pipeline/rds_sync_detector.py
  - Streaming RDS BPSK chain: -57 kHz freq shift -> LPF -> /10 -> M&M
    (cubic interpolation) -> Costas loop -> SyncEvent
  - Independent pilot extraction for CrystalCalibrator
  - Soft gap handling preserves M&M timing across freq_hop blocks
  - Audio normalised to unit RMS so M&M loop gain works regardless of FM
    deviation amplitude

SyncEvent.sample_index: int -> float to preserve M&M sub-sample precision.
Verified 0.01 usec interval jitter on synthetic data (was 1.93 usec when
rounded to int, was thousands of usec with pilot zero-crossing ambiguity).

Pipeline integration:
  - PipelineConfig.sync_mode: defaults to "rds" (only mode wired in;
    extensibility hook for future detectors)
  - Server _T_SYNC_NS: 7,000,000 ns -> 1e9/1187.5 ns (842 usec RDS bit period)
  - Disambiguation tests rewritten for the new sync period
  - Mismatched sync_tx test: now verifies RDS is robust to ~99 usec
    coordinate errors (was an n=wrong failure with the old 7 ms period)

Phase 3 prep: scripts/verify_rds_sync.py — live diagnostic mirroring
verify_sync.py but with RDS event rate, pilot quality, crystal drift,
and bit interval jitter checks.

Tests: 628 passed (17 new RDS unit tests, e2e tests updated to generate
synthetic FM IQ with pilot + 57 kHz RDS BPSK).

+++ src/beagle_node/pipeline/delta.py
@@ -76,7 +76,7 @@ class TDOAMeasurement:
     """
     sync_delta_ns: int                  # THE measurement
     target_sample: int                  # triggering event sample index
-    sync_sample: int                    # SyncEvent sample index used
+    sync_sample: float                  # SyncEvent sample index used (sub-sample precision)
     sample_rate_hz: float               # Corrected sample rate
     sample_rate_correction: float       # CrystalCalibrator factor
     pps_anchored: bool                  # True if GPS 1PPS was used
